refactor: Page|Paginatin|Limit|PaginationResults primitive wrapper type change to primitive type

+++ common-core/src/main/java/cn/featherfly/common/structure/page/Pagination.java
@@ -12,14 +12,14 @@ public interface Pagination extends Page {
      *
      * @return 总页数
      */
-    Integer getTotalPage();
+    long getTotalPage();
 
     /**
      * 获取总数.
      *
      * @return 总数
      */
-    Integer getTotal();
+    long getTotal();
 
     //    /**
     //     * <p>
